just grow bjs.conf by one node.
bjs is mostly working.
bpmaster has a problem in initialization. result is that if there are no nodes
and you do this
bpsh 0 date

bpmaster does this:
send_msg:2898: slave 0 is not running

Program received signal SIGABRT, Aborted.
0x00007ffff7aa34b5 in raise () from /lib/libc.so.6

it's because the initial value of the slave state is wrong. need to fix.